+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Next
 - Fix: only use sticky highlight when using tooltip in SVGMap
+- Fix: update SVGMap's internal getHighlightFeature to work with new interaction functionality
 
 ## v2.0.0